+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Anticipated release: December 16, 2019
 - Upload files from the rich text editor instead of embedding them directly into the text ([#1735])
 - Made the rich text areas resize to fit their contents
 - Automatically save APDs as changes are made
+- Removed the 11 Standards and Conditions text fields in favor of two simpler fields ([#1947])
 
 #### 🐛 Bugs fixed
 
@@ -213,6 +214,7 @@ Pilot release to select state partners
 [#1921]: https://github.com/18F/cms-hitech-apd/pull/1921
 [#1914]: https://github.com/18F/cms-hitech-apd/issues/1914
 [#1925]: https://github.com/18F/cms-hitech-apd/issues/1925
+[#1928]: https://github.com/18F/cms-hitech-apd/issues/1928
+[#1947]: https://github.com/18F/cms-hitech-apd/issues/1947
 [#1957]: https://github.com/18F/cms-hitech-apd/pull/1957
 [#1967]: https://github.com/18F/cms-hitech-apd/pull/1967
-[#1928]: https://github.com/18F/cms-hitech-apd/issues/1928